third_party/blink/web_tests/platform/generic/fast/events/event-sender-mouseleave-expected.txt

PASS hasMouseLeave is true
PASS hasMouseOut is true
PASS box.offsetLeft + 10 is lastMouseX
PASS box.offsetTop + 10 is lastMouseY
PASS successfullyParsed is true

TEST COMPLETE


Test that a mouse leave event is triggered when the OS tells us the mouse has left without changing position.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".
